pyproject.toml
[tool.poetry]
name = "reuse-agent"
version = "0.1.0"
description = "reuse agent"
authors = ["Bird <aflybird0@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.3.7"
langchain = "^0.0.347"
tree-of-thoughts = "^0.3.7"
numexpr = "^2.8.7"
langchainhub = "^0.1.14"
langchain-experimental = "^0.0.44"
pymongo = "^4.6.1"
langserve = "^0.0.34"
fastapi = "^0.104.1"
sse-starlette = "^1.8.2"
pydantic = "1.10.13"
streamlit = "^1.29.0"
watchdog = "^3.0.0"
streamlit-tags = "^1.2.8"
st-annotated-text = "^4.0.1"
uvicorn = "^0.25.0"
docker = "^7.0.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
